requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
greenlet==3.0.1
idna==3.4
image==1.5.33
Pillow==10.0.0
playwright==1.37.0
pyee==11.0.1
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.7
